Switched to stable norm calculation.
The implementation of the L2 norm seems to be numerically more robust in
recent versions of Eigen. In order to avoid errors with older versions
of Eigen, the stableNorm() routine is now used at critical points in the
calculation. It might be worth to experiment further in this direction,
also with regards to the performance of the different methods of
computing the L2 norm.