Preparations for calculation of sphere area.
In addition to the calculation of the surface area of part of the
element contained within the sphere, we also want to determine the
respective area of the sphere. This commit modifies the interface for
the area calculation to return the surface are of the part of the sphere
within the element as the first element in the result array.