Introduce namespace 'overlap'
The library now uses a dedicated namespace 'overlap' to avoid potential
name conflicts when including the library in other projects. Only the
types and methods forming the public API are exposed in this namespace,
everything else is contained in the namespace 'overlap::detail'.
Also, the free functions to calculate the overlap volume and areas are
renamed to 'overlap_volume' and 'overlap_area', respectively.